Return the value of the fuzz factor used in numeric expressions.

Return the value of the fuzz factor used in numeric expressions.

Usage

config_epsilon()

Value

A float.

Examples

config_epsilon()

## [1] 1e-07

See also