Returns the list of input tensors necessary to compute tensor.

Output will always be a list of tensors (potentially with 1 element).

Usage

get_source_inputs(tensor)

Arguments

tensor

The tensor to start from.

Value

List of input tensors.

Example

input <- keras_input(c(3))
output <- input |> layer_dense(4) |> op_multiply(5)
reticulate::py_id(get_source_inputs(output)[[1]]) ==
reticulate::py_id(input)

## [1] TRUE

See also