keras3: R Interface to 'Keras'

Interface to 'Keras' https://keras.io, a high-level neural networks 'API'. 'Keras' was developed with a focus on enabling fast experimentation, supports both convolution based networks and recurrent networks (as well as combinations of the two), and runs seamlessly on both 'CPU' and 'GPU' devices.

Keras is a high-level neural networks API, developed with a focus on enabling fast experimentation. Keras has the following key features:

Details

Allows the same code to run on CPU or on GPU, seamlessly.

User-friendly API which makes it easy to quickly prototype deep learning models.

Built-in support for convolutional networks (for computer vision), recurrent networks (for sequence processing), and any combination of both.

Supports arbitrary network architectures: multi-input or multi-output models, layer sharing, model sharing, etc. This means that Keras is appropriate for building essentially any deep learning model, from a memory network to a neural Turing machine.

Is capable of running on top of multiple back-ends including TensorFlow, Jax, or PyTorch.

See the package website at https://keras.posit.co for complete documentation.

See also

Author

Maintainer: Tomasz Kalinowski tomasz@posit.co [copyright holder]

Authors:

JJ Allaire [copyright holder]

François Chollet [copyright holder]

Other contributors:

Daniel Falbel daniel@posit.co [contributor, copyright holder]

Posit Software, PBC [copyright holder, funder]

Google [copyright holder, funder]

Yuan Tang terrytangyuan@gmail.com (ORCID) [contributor, copyright holder]

Wouter Van Der Bijl [contributor, copyright holder]

Martin Studer [contributor, copyright holder]

Sigrid Keydana [contributor]