3D transposed convolution layer.

The need for transposed convolutions generally arise from the desire to use a transformation going in the opposite direction of a normal convolution, i.e., from something that has the shape of the output of some convolution to something that has the shape of its input while maintaining a connectivity pattern that is compatible with said convolution.

Usage

layer_conv_3d_transpose(
  object,
  filters,
  kernel_size,
  strides = list(1L, 1L, 1L),
  padding = "valid",
  data_format = NULL,
  dilation_rate = list(1L, 1L, 1L),
  activation = NULL,
  use_bias = TRUE,
  kernel_initializer = "glorot_uniform",
  bias_initializer = "zeros",
  kernel_regularizer = NULL,
  bias_regularizer = NULL,
  activity_regularizer = NULL,
  kernel_constraint = NULL,
  bias_constraint = NULL,
  ...
)

Arguments

object

Object to compose the layer with. A tensor, array, or sequential model.

filters

int, the dimension of the output space (the number of filters in the transposed convolution).

kernel_size

int or list of 1 integer, specifying the size of the transposed convolution window.

strides

int or list of 1 integer, specifying the stride length of the transposed convolution. strides > 1 is incompatible with dilation_rate > 1.

padding

string, either "valid" or "same" (case-insensitive). "valid" means no padding. "same" results in padding evenly to the left/right or up/down of the input. When padding="same" and strides=1, the output has the same size as the input.

data_format

string, either "channels_last" or "channels_first". The ordering of the dimensions in the inputs. "channels_last" corresponds to inputs with shape (batch_size, spatial_dim1, spatial_dim2, spatial_dim3, channels) while "channels_first" corresponds to inputs with shape (batch_size, channels, spatial_dim1, spatial_dim2, spatial_dim3). It defaults to the image_data_format value found in your Keras config file at ~/.keras/keras.json. If you never set it, then it will be "channels_last".

dilation_rate

int or list of 1 integers, specifying the dilation rate to use for dilated transposed convolution.

activation

Activation function. If NULL, no activation is applied.

use_bias

bool, if TRUE, bias will be added to the output.

kernel_initializer

Initializer for the convolution kernel. If NULL, the default initializer ("glorot_uniform") will be used.

bias_initializer

Initializer for the bias vector. If NULL, the default initializer ("zeros") will be used.

kernel_regularizer

Optional regularizer for the convolution kernel.

bias_regularizer

Optional regularizer for the bias vector.

activity_regularizer

Optional regularizer function for the output.

kernel_constraint

Optional projection function to be applied to the kernel after being updated by an Optimizer (e.g. used to implement norm constraints or value constraints for layer weights). The function must take as input the unprojected variable and must return the projected variable (which must have the same shape). Constraints are not safe to use when doing asynchronous distributed training.

bias_constraint

Optional projection function to be applied to the bias after being updated by an Optimizer.

...

For forward/backward compatability.

Value

A 5D tensor representing activation(conv3d(inputs, kernel) + bias).

Input Shape

If data_format="channels_last": 5D tensor with shape: (batch_size, spatial_dim1, spatial_dim2, spatial_dim3, channels)

If data_format="channels_first": 5D tensor with shape: (batch_size, channels, spatial_dim1, spatial_dim2, spatial_dim3)

Output Shape

If data_format="channels_last": 5D tensor with shape: (batch_size, new_spatial_dim1, new_spatial_dim2, new_spatial_dim3, filters)

If data_format="channels_first": 5D tensor with shape: (batch_size, filters, new_spatial_dim1, new_spatial_dim2, new_spatial_dim3)

Raises

ValueError: when both strides > 1 and dilation_rate > 1.

References

A guide to convolution arithmetic for deep learning

Deconvolutional Networks

Example

x <- random_uniform(c(4, 10, 8, 12, 128))
y <- x |> layer_conv_3d_transpose(32, 2, 2, activation = 'relu')
shape(y)

## shape(4, 20, 16, 24, 32)

See also