Wraps arbitrary expressions as a Layer object.

The layer_lambda() layer exists so that arbitrary expressions can be used as a Layer when constructing Sequential and Functional API models. Lambda layers are best suited for simple operations or quick experimentation. For more advanced use cases, prefer writing new subclasses of Layer using new_layer_class().

Usage

layer_lambda(
  object,
  f,
  output_shape = NULL,
  mask = NULL,
  arguments = NULL,
  ...
)

Arguments

object

Object to compose the layer with. A tensor, array, or sequential model.

f

The function to be evaluated. Takes input tensor as first argument.

output_shape

Expected output shape from function. This argument can usually be inferred if not explicitly provided. Can be a list or function. If a list, it only specifies the first dimension onward; sample dimension is assumed either the same as the input: output_shape = c(input_shape[1], output_shape) or, the input is NULL and the sample dimension is also NULL: output_shape = c(NA, output_shape). If a function, it specifies the entire shape as a function of the input shape: output_shape = f(input_shape).

mask

Either NULL (indicating no masking) or a callable with the same signature as the compute_mask layer method, or a tensor that will be returned as output mask regardless of what the input is.

arguments

Optional named list of arguments to be passed to the function.

...

For forward/backward compatability.

Value

The return value depends on the value provided for the first argument. If object is:

a keras_model_sequential(), then the layer is added to the sequential model (which is modified in place). To enable piping, the sequential model is also returned, invisibly.

a keras_input(), then the output tensor from calling layer(input) is returned.

NULL or missing, then a Layer instance is returned.

Examples

# add a x -> x^2 layer
model <- keras_model_sequential()
model |> layer_lambda(\(x) x^2)

See also