Reduces the number of bits for each color channel.

Reduces the number of bits for each color channel.

Usage

layer_random_posterization(
  object,
  factor,
  value_range = list(0L, 255L),
  data_format = NULL,
  seed = NULL,
  ...
)

Arguments

object

Object to compose the layer with. A tensor, array, or sequential model.

factor

integer, the number of bits to keep for each channel. Must be a value between 1-8.

value_range

a tuple or a list of two elements. The first value represents the lower bound for values in passed images, the second represents the upper bound. Images passed to the layer should have values within value_range. Defaults to (0, 255).

data_format

String, one of "channels_last" (default) or "channels_first". The ordering of the dimensions in the inputs. "channels_last" corresponds to inputs with shape (batch, height, width, channels) while "channels_first" corresponds to inputs with shape (batch, channels, height, width).

seed

Initial seed for the random number generator

...

For forward/backward compatability.

References

AutoAugment: Learning Augmentation Policies from Data

RandAugment: Practical automated data augmentation with a reduced search space

See also