Unit normalization layer.

Normalize a batch of inputs so that each input in the batch has a L2 norm equal to 1 (across the axes specified in axis).

Usage

layer_unit_normalization(object, axis = -1L, ...)

Arguments

object

Object to compose the layer with. A tensor, array, or sequential model.

axis

Integer or list. The axis or axes to normalize across. Typically, this is the features axis or axes. The left-out axes are typically the batch axis or axes. -1 is the last dimension in the input. Defaults to -1.

...

For forward/backward compatability.

Value

The return value depends on the value provided for the first argument. If object is:

a keras_model_sequential(), then the layer is added to the sequential model (which is modified in place). To enable piping, the sequential model is also returned, invisibly.

a keras_input(), then the output tensor from calling layer(input) is returned.

NULL or missing, then a Layer instance is returned.

Examples

data <- op_reshape(1:6, newshape = c(2, 3))
normalized_data <- layer_unit_normalization(data)
op_sum(normalized_data[1,]^2)

## tf.Tensor(0.9999999, shape=(), dtype=float32)

See also