Define new keras types

These functions can be used to make custom objects that fit in the family of existing keras types. For example, new_layer_class() will return a class constructor, an object that behaves like other layer functions such as layer_dense(). new_callback_class() will return an object that behaves similarly to other callback functions, like callback_reduce_lr_on_plateau(), and so on. All arguments with a default NULL value are optional methods that can be provided.

Usage

new_metric_class(classname, ..., initialize, update_state, result)

new_loss_class(
  classname,
  ...,
  call = NULL,
  inherit = keras3::keras$losses$Loss
)

new_model_class(
  classname,
  ...,
  initialize = NULL,
  call = NULL,
  train_step = NULL,
  predict_step = NULL,
  test_step = NULL,
  compute_loss = NULL,
  compute_metrics = NULL
)

Arguments

classname

The classname as a string. Convention is for the classname to be a CamelCase version of the constructor.

...

Additional fields and methods for the new type.

initialize, update_state, result, call, train_step, predict_step, test_step, compute_loss, compute_metrics

Optional methods that can be overridden.

inherit

class to inherit from.

Value

A new class generator object that inherits from the appropriate Keras base class.

Details

mark_active() is a decorator that can be used to indicate functions that should become active properties of the class instances.