Gives a new shape to a tensor without changing its data.

Gives a new shape to a tensor without changing its data.

Usage

op_reshape(x, newshape)

Arguments

x

Input tensor.

newshape

The new shape should be compatible with the original shape. One shape dimension can be -1 in which case the value is inferred from the length of the array and remaining dimensions.

Value

The reshaped tensor.

See also