Draw samples from a uniform distribution.

The generated values follow a uniform distribution in the range [minval, maxval). The lower bound minval is included in the range, while the upper bound maxval is excluded.

dtype must be a floating point type, the default range is [0, 1).

Usage

random_uniform(shape, minval = 0, maxval = 1, dtype = NULL, seed = NULL)

Arguments

shape

The shape of the random values to generate.

minval

Float, defaults to 0. Lower bound of the range of random values to generate (inclusive).

maxval

Float, defaults to 1. Upper bound of the range of random values to generate (exclusive).

dtype

Optional dtype of the tensor. Only floating point types are supported. If not specified, config_floatx() is used, which defaults to float32 unless you configured it otherwise (via config_set_floatx(float_dtype))

seed

An R integer or instance of random_seed_generator(). Used to make the behavior of the initializer deterministic. Note that an initializer seeded with an integer or NULL (unseeded) will produce the same random values across multiple calls. To get different random values across multiple calls, use as seed an instance of random_seed_generator().

See also