Saves all layer weights to a .weights.h5 file.

Saves all layer weights to a .weights.h5 file.

Usage

save_model_weights(model, filepath, overwrite = FALSE)

Arguments

model

A keras Model object

filepath

string. Path where to save the model. Must end in .weights.h5.

overwrite

Whether we should overwrite any existing model at the target location, or instead ask the user via an interactive prompt.

Value

This is called primarily for side effects. model is returned, invisibly, to enable usage with the pipe.

See also