Tensor shape utility

This function can be used to create or get the shape of an object.

Usage

shape(...)

# S3 method for keras_shape
format(x, ...)

# S3 method for keras_shape
print(x, ...)

# S3 method for keras_shape
[(x, ...)

# S3 method for keras_shape
r_to_py(x, convert = FALSE)

# S3 method for keras_shape
as.integer(x, ...)

# S3 method for keras_shape
as.list(x)

Arguments

...

A shape specification. Numerics, NULL and tensors are valid. NULL, NA, and -1L can be used to specify an unspecified dim size. Tensors are dispatched to k_shape() to extract the tensor shape. Values wrapped in I() are used asis (see examples). All other objects are coerced via as.integer().

Value

A list with a "keras_shape" class attribute. Each element of the list will be either a) NULL, b) an integer or c) a scalar integer tensor (e.g., when supplied a TF tensor with a unspecified dimension in a function being traced).

Examples

shape(1, 2, 3)

## shape(1, 2, 3)

3 ways to specify an unknown dimension

shape(NA,   2, 3)
shape(NULL, 2, 3)
shape(-1,   2, 3)

## shape(NA, 2, 3)
## shape(NA, 2, 3)
## shape(NA, 2, 3)

Most functions that take a 'shape' argument also coerce with shape()

layer_input(c(1, 2, 3))
layer_input(shape(1, 2, 3))

## <KerasTensor shape=(None, 1, 2, 3), dtype=float32, sparse=None, name=keras_tensor>
## <KerasTensor shape=(None, 1, 2, 3), dtype=float32, sparse=None, name=keras_tensor_1>

You can also use shape() to get the shape of a tensor

symbolic_tensor <- layer_input(shape(1, 2, 3))
shape(symbolic_tensor)

## shape(NA, 1, 2, 3)

eager_tensor <- op_ones(c(1,2,3))
shape(eager_tensor)

## shape(1, 2, 3)

op_shape(eager_tensor)

## shape(1, 2, 3)

Combine or expand shapes

shape(symbolic_tensor, 4)

## shape(NA, 1, 2, 3, 4)

shape(5, symbolic_tensor, 4)

## shape(5, NA, 1, 2, 3, 4)

In graph mode, a shape might contain a scalar integer tensor for unknown axes.

tfn <- tensorflow::tf_function(function(x) {
  print(shape(x))
  x
},
input_signature = list(tensorflow::tf$TensorSpec(shape(1, NA, 3))))
invisible(tfn(op_ones(shape(1, 2, 3))))

## shape(1, Tensor("strided_slice:0", shape=(), dtype=int32), 3)

A useful pattern is to unpack the shape() with %<-%, like this:

c(batch_size, seq_len, channels) %<-% shape(x)

If you are unpacking shape() in graph mode, and then want to reassemble the axes with shape(), you'll have to wrap tensors with I() to use the tensor itself, rather than the shape of the tensor.

echo_print <- function(x) { message("> ", deparse(substitute(x))); print(x) }
tfn <- tensorflow::tf_function(function(x) {
  c(axis1, axis2, axis3) %<-% shape(x)
  str(list(axis1 = axis1, axis2 = axis2, axis3 = axis3))

  echo_print(shape(axis2))               # resolve axis2 tensor shape
  echo_print(shape(axis1, axis2, axis3)) # resolve axis2 tensor shape

  echo_print(shape(I(axis2)))               # use axis2 tensor as axis value
  echo_print(shape(axis1, I(axis2), axis3)) # use axis2 tensor as axis value
  x
},
input_signature = list(tensorflow::tf$TensorSpec(shape(1, NA, 3))))
invisible(tfn(op_ones(shape(1, 2, 3))))

## List of 3
##  $ axis1: int 1
##  $ axis2:<tf.Tensor 'strided_slice:0' shape=() dtype=int32>
##  $ axis3: int 3

## > shape(axis2)

## shape()

## > shape(axis1, axis2, axis3)

## shape(1, 3)

## > shape(I(axis2))

## shape(Tensor("strided_slice:0", shape=(), dtype=int32))

## > shape(axis1, I(axis2), axis3)

## shape(1, Tensor("strided_slice:0", shape=(), dtype=int32), 3)

See also